EFFECTIVE MOBILE - Tasks/Pytest/Task/requirements.txt
asyncpg
SQLAlchemy~=2.0.40
fastapi~=0.115.12
uvicorn[standard]
python-dotenv~=1.1.0
pydantic~=2.11.4
redis>=4.5.0